Migrate origin/main's new code to zig 0.16 (compiles)
`zig build` now compiles the full compiler + all artifacts under zig 0.16.
Migrated the 0.15.2 code that origin/main added (which the zig-16 branch had
not yet seen) using the established patterns:

- Layout `.data.<field>` packed-int access → accessor methods
  (getScalar/getStruct/getTagUnion/getClosure/getIdx; Scalar.getInt/getFrac).
- std.ArrayList / *HashMapUnmanaged empty init `.{}` → `.empty`.
- Removed std APIs: posix.write/exit/close/waitpid/fork/pipe/getenv,
  process.getEnvVarOwned, process.Child.run/init, time.Timer/milliTimestamp,
  crypto.random, fs.cwd, Thread.Mutex/sleep → 0.16 equivalents
  (std.c.*, harness helpers, std.Io.*, std.process.run/spawn, Environ,
  std.atomic.Mutex, posix monotonic clock).
- Thread io through SharedMemoryAllocator.createWithMinSize, the
  test_helpers lowering chain, and llvm_compile (compileToObject etc.).
- signal_handler: io-free atomic mutex; sigaction handler takes posix.SIG;
  stack-pointer-from-context returns null (classifyFault falls back to the
  faulting address); ported the stack_overflow test to std.process.run.
- Bump CI workflow zig to 0.16.0.

Next: `zig build minici` (tests, lints, snapshots).

+++ src/compile/test/module_env_test.zig
@@ -120,7 +120,7 @@ test "ModuleEnv.Serialized roundtrip" {
     try testing.expectEqual(@as(u32, 85), env.common.idents.interner.entry_count);
 
     try testing.expectEqual(@as(usize, 1), env.common.exposed_items.count());
-    try testing.expectEqual(@as(?u16, 42), env.common.exposed_items.getNodeIndexById(gpa, @as(u32, @bitCast(hello_idx))));
+    try testing.expectEqual(@as(?u32, 42), env.common.exposed_items.getNodeIndexById(gpa, @as(u32, @bitCast(hello_idx))));
 
     try testing.expectEqual(@as(usize, 3), env.common.line_starts.len());
     try testing.expectEqual(@as(u32, 0), env.common.line_starts.items.items[0]);
